ref(server): Rename module to actors::envelopes
