+++ relay-server/src/actors/mod.rs
@@ -13,10 +13,10 @@
 //!  - [`ProjectCache`](project_cache::ProjectCache): A cache that serves queries for project
 //!    configurations. Its requests are debounced and batched based on a configured interval (100ms
 //!    by default). Also, missing projects are cached for some time.
-//!  - [`EventManager`](events::EventManager) and [`EventProcessor`](events::EventProcessor): Handle
-//!    a queue of events, verify their projects, execute PII stripping and finally send the event to
-//!    the upstream. The processor is spawned in multiple synchronous worker threads (via
-//!    `SyncArbiter`).
+//!  - [`EventManager`](envelopes::EventManager) and [`EventProcessor`](envelopes::EventProcessor):
+//!    Handle a queue of envelopes, verify their projects, execute PII stripping and finally send
+//!    the envelope to the upstream. The processor is spawned in multiple synchronous worker threads
+//!    (via `SyncArbiter`).
 //!  - [`UpstreamRelay`](upstream::UpstreamRelay): Abstraction for communication with the upstream
 //!    (either another Relay or Sentry). It manages an internal client connector to throttle
 //!    requests and ensures this relay is authenticated before sending queries (e.g. project config
@@ -34,7 +34,7 @@
 
 pub mod connector;
 pub mod controller;
-pub mod events;
+pub mod envelopes;
 pub mod healthcheck;
 pub mod outcome;
 pub mod project;
